Add handling SQLiteDatabaseCorruptException.
When SQLiteDatabaseCorruptException occurred on execute() or executeForString(),
DefaultDatabaseErrorHandler.onCorrupt() is not to be called.
So, Malformed DB file can't be deleted.
Because, Instead SQLiteDatabaseCorruptException, code is throwing SQLiteException.
On boot time, system can't open DB file and it throws RuntimeException.
This causes infinite reset.
So, I modified to throw SQLiteDatabaseCorruptException.

	modified:   core/java/android/database/sqlite/SQLiteConnection.java

Change-Id: I125b69d62acc4c0690b8dd17d4e14b49be141095